Corrections des tests unitaires et validation
